# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Log Relay Project.",
        "title": "Log Relay API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/user/register": {
            "post": {
                "description": "Creates a user account, hashes the password, and hashes and API token.\nReturns a one-time API token and sets an HttpOnly 'token' cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Register New User",
                "parameters": [
                    {
                        "description": "User Registration Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_auth.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_auth.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/codewax_internal_dtos.ValidationErrorResponse"
                        }
                    },
                    "409": {
                        "description": "User already exists with email",
                        "schema": {
                            "$ref": "#/definitions/codewax_internal_dtos.AlreadyExistsResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/codewax_internal_dtos.ServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "codewax_internal_dtos.AlreadyExistsResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Already in use"
                }
            }
        },
        "codewax_internal_dtos.ServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "An unexpected error occurred"
                }
            }
        },
        "codewax_internal_dtos.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "Validation failed"
                }
            }
        },
        "internal_handlers_auth.RegisterInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "User Name"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8,
                    "example": "SecurePass123!"
                }
            }
        },
        "internal_handlers_auth.RegisterResponse": {
            "type": "object",
            "properties": {
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "Registration successful"
                },
                "user_email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "user_name": {
                    "type": "string",
                    "example": "User Name"
                }
            }
        }
    }
}